% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DA.ci.R
\name{plot.DA_ci}
\alias{plot.DA_ci}
\title{plot method for DA_ci objects}
\usage{
\method{plot}{DA_ci}(x, xlab = "Time", ylab = "Cumulative incidence", ...)
}
\arguments{
\item{x}{A DA_ci object}

\item{xlab}{Label for x axis}

\item{ylab}{Label for y axix}

\item{...}{other arguments}
}
\description{
plot method for DA_ci objects
}

