% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DA.surv.R
\name{plot.DA_surv}
\alias{plot.DA_surv}
\title{plot method for DA_surv objects}
\usage{
\method{plot}{DA_surv}(x, xlab = "Time", ylab = "Survival",
  fun = c("surv", "event"), ...)
}
\arguments{
\item{x}{a DA_surv object}

\item{xlab}{Label for x axis}

\item{ylab}{Label for y axix}

\item{fun}{If fun = event, 1 - the surv is drown}

\item{...}{other arguments}
}
\description{
plot method for DA_surv objects
}

