% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.MI_ci}
\alias{plot.MI_ci}
\title{plot method for MI_ci objects}
\usage{
\method{plot}{MI_ci}(x, xlab = "Time", ylab = "Cumulative incidence", ...)
}
\arguments{
\item{x}{A MI_ci object}

\item{xlab}{Label for x axis}

\item{ylab}{Label for y axix}

\item{...}{other arguments}
}
\description{
plot method for MI_ci objects
}

