% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.MI_surv}
\alias{plot.MI_surv}
\title{plot method for MI_surv objects}
\usage{
\method{plot}{MI_surv}(x, xlab = "Time", ylab = "Survival",
  fun = c("surv", "event"), ...)
}
\arguments{
\item{x}{a MI_surv object}

\item{xlab}{Label for x axis}

\item{ylab}{Label for y axix}

\item{fun}{If fun = event, 1 - the surv is drown}

\item{...}{other arguments}
}
\description{
plot method for MI_surv objects
}

