% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MIICD.coxph.R
\name{plot.MIICD_coxph}
\alias{plot.MIICD_coxph}
\title{plot method for MIICD_coxph objects}
\usage{
\method{plot}{MIICD_coxph}(x, type = c("baseline", "coef", "sigma"),
  coef = 1, ylab = "Survival", xlab = "Time", ...)
}
\arguments{
\item{x}{a MIICD_coxph object}

\item{type}{type of diagnoctic plot to display}

\item{coef}{An integer: the no of the coefficient to display}

\item{ylab}{Label for y axix}

\item{xlab}{Label for x axis}

\item{...}{other arguments}
}
\description{
plot method for MIICD_coxph objects
}

