% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MIICD.crreg.R
\name{plot.MIICD_crreg}
\alias{plot.MIICD_crreg}
\title{plot method for MIICD_crreg objects}
\usage{
\method{plot}{MIICD_crreg}(x, type = c("baseline", "coef", "sigma"),
  coef = 1, ylab = "Cumulative incidence", xlab = "Time", ...)
}
\arguments{
\item{x}{a MIICD_crreg object}

\item{type}{type of diagnoctic plot to display}

\item{coef}{An integer: the no of the coefficient to display}

\item{ylab}{Label for y axix}

\item{xlab}{Label for x axis}

\item{...}{other arguments}
}
\description{
plot method for MIICD_crreg objects
}

