% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{bcos}
\alias{bcos}
\title{bcos : breast cosmesis data}
\description{
A data frame with 94 observations on the following 3 variables : \itemize{
\item left
\item right
\item treatment
}
The data comes from the \code{Interval} library by Michael P. Fay.
}
\examples{
head(bcos)
}
\author{
Michael P. Fay
}
\references{
Finkelstein, D.M., and Wolfe, R.A. (1985). A semiparametric model for regression analysis of interval-censored failure time data. Biometrics 41: 731-740.
}
\keyword{dataset}

